feat: 🎸 build library without including less-styling
we're moving away from less as a technology, and don't want to force it
on anyone. The "nav-frontend-X" packages are all exporting a precompiled
css to help in this. And therefore so should this library

BREAKING CHANGE: 🧨 Users must ensure that all styling from nav-frontend is included in
their app after updating